+++ Configuration/TCA/tx_filetransfer_domain_model_upload.php
@@ -8,13 +8,23 @@ return [
         'label' => 'subject',
         'tstamp' => 'tstamp',
         'crdate' => 'crdate',
-        'versioningWS' => true,
-        'delete' => 'deleted',
+        'versioningWS' => false,
         'iconfile' => 'EXT:filetransfer/Resources/Public/Icons/Extension.gif',
-        'origUid' => 't3_origuid',
-        'hideTable' => true,
+        'origUid' => 't3_origuid'
     ],
     'columns' => [
+        'sender_address' => [
+            'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.sender_address',
+            'config' => [
+                'type' => 'email',
+            ],
+        ],
+        'receiver_address' => [
+            'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.receiver_address',
+            'config' => [
+                'type' => 'email',
+            ],
+        ],
         'subject' => [
             'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.subject',
             'config' => [
@@ -43,14 +53,12 @@ return [
                 'max' => 80,
             ],
         ],
-        'downloaded' => [
-            'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.downloaded',
+        'download_limit' => [
+            'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.download_limit',
             'config' => [
                 'type' => 'number',
+                'format' => 'integer',
                 'readonly' => true,
-                'range' => [
-                    'lower' => 0,
-                ],
             ],
         ],
         'asset' => [
@@ -65,10 +73,22 @@ return [
             'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.validity_date',
             'config' => [
                 'type' => 'datetime',
+                'readonly' => true,
+            ],
+        ],
+        'validity_duration' => [
+            'label' => 'LLL:EXT:filetransfer/Resources/Private/Language/locallang_db.xlf:tx_filetransfer_domain_model_upload.validity_date',
+            'config' => [
+                'type' => 'number',
+                'format' => 'integer',
+                'default' => 3,
+                'range' => [
+                    'lower' => 3,
+                ]
             ],
         ],
     ],
     'types' => [
-        0 => ['showitem' => 'subject,message,asset,validity_date,downloaded,token'],
+        0 => ['showitem' => 'sender_address,receiver_address,subject,message,asset,validity_duration,validity_date,downloaded,token'],
     ],
 ];