+++ ext_emconf.php
@@ -18,4 +18,9 @@ $EM_CONF[$_EXTKEY] = [
         'suggests' => [
         ],
     ],
+    'autoload' => [
+        'psr-4' => [
+            'Wacon\\Filetransfer\\' => 'Classes',
+        ]
+    ],
 ];