+++ CHANGES.md
@@ -1,3 +1,11 @@
+## Version 0.0.3
+- [IMPORTANT] Rename downloaded to download_limit
+- [FEATURE] Add Download Plugin
+- [FEATURE] Add receiver and sender address
+- [FEATURE] Add Upload Form Template
+- [CHANGE] Remove versioning and deleted flag for Upload Model
+- [CHANGE] Ignore .vscode from git
+
 ## Version 0.0.2
 - [FEATURE] Create Upload Model
 - [FEATURE] Register Typoscript